pyproject.toml
[project]
name = "sussia"
version = "0.1.0"
description = "A turn-based multiplayer strategy game"
authors = [
    {name = "Andrei",email = "andreisva2023@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib (>=3.10.0,<4.0.0)",
    "discord (>=2.3.2,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "numpy (<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.394"

[tool.poetry]
package-mode = true